+++ src/dtkCore/dtkCoreParameter.h
@@ -102,12 +102,15 @@ public:
     virtual QVariant variant(void) const = 0;
     virtual QVariantHash toVariantHash(void) const = 0;
 
+    dtkCoreParameter *castToTop(void);
+
 #pragma mark - Connection management
 
     void block(bool);
     void sync(void);
     template <typename F> QMetaObject::Connection connect(F);
     void disconnect(void);
+    void disconnect(QMetaObject::Connection);
 
     void syncFail(void);
     template <typename F> QMetaObject::Connection connectFail(F);
@@ -172,6 +175,7 @@ public:
     using self_type = dtkCoreParameterSimple<value_type>;
     using base_type = dtkCoreParameterBase<self_type>;
 
+    using dtkCoreParameter::castToTop;
     using dtkCoreParameter::documentation;
     using dtkCoreParameter::setDocumentation;
     using dtkCoreParameter::label;
@@ -228,6 +232,7 @@ public:
     using self_type = dtkCoreParameterNumeric<value_type>;
     using base_type = dtkCoreParameterBase<self_type>;
 
+    using dtkCoreParameter::castToTop;
     using dtkCoreParameter::documentation;
     using dtkCoreParameter::setDocumentation;
     using dtkCoreParameter::label;
@@ -338,6 +343,7 @@ public:
     using self_type = dtkCoreParameterInList<value_type>;
     using base_type = dtkCoreParameterBase<self_type>;
 
+    using dtkCoreParameter::castToTop;
     using dtkCoreParameter::documentation;
     using dtkCoreParameter::setDocumentation;
     using dtkCoreParameter::label;
@@ -389,6 +395,64 @@ DTKCORE_EXPORT QDataStream& operator >> (QDataStream&, dtkCoreParameterInList<T>
 template <typename T>
 DTKCORE_EXPORT QDebug operator << (QDebug, dtkCoreParameterInList<T>);
 
+
+// ///////////////////////////////////////////////////////////////////
+// dtkCoreParameter contained in a given list
+// ///////////////////////////////////////////////////////////////////
+
+class DTKCORE_EXPORT dtkCoreParameterInListStringList : public dtkCoreParameterBase<dtkCoreParameterInListStringList>
+{
+public:
+    using value_type = QStringList;
+    using self_type = dtkCoreParameterInListStringList;
+    using base_type = dtkCoreParameterBase<self_type>;
+
+    using dtkCoreParameter::castToTop;
+    using dtkCoreParameter::documentation;
+    using dtkCoreParameter::setDocumentation;
+    using dtkCoreParameter::label;
+    using dtkCoreParameter::setLabel;
+
+public:
+     dtkCoreParameterInListStringList(void) = default;
+    ~dtkCoreParameterInListStringList(void) = default;
+
+    dtkCoreParameterInListStringList(const QStringList&);
+    dtkCoreParameterInListStringList(const QVariant&);
+    dtkCoreParameterInListStringList(const dtkCoreParameterInListStringList&);
+
+    dtkCoreParameterInListStringList(const QString&, const QStringList&, const QStringList&, const QString& = QString());
+
+    dtkCoreParameterInListStringList& operator = (const QStringList&);
+    dtkCoreParameterInListStringList& operator = (const QVariant&);
+    dtkCoreParameterInListStringList& operator = (const dtkCoreParameterInListStringList&);
+
+    int size(void) const;
+    QStringList value(void) const;
+    QStringList values(void) const;
+
+    void addValue(const QString&);
+    void removeValue(const QString&);
+
+    void setValues(const QStringList&);
+    void setValue(const QStringList&);
+    void setValue(const QVariant&);
+    QVariantHash toVariantHash(void) const override;
+
+private:
+    using dtkCoreParameter::m_label;
+    using dtkCoreParameter::m_doc;
+
+    QStringList m_value;
+    QStringList m_values;
+};
+
+DTKCORE_EXPORT QDataStream& operator << (QDataStream&, const dtkCoreParameterInListStringList&);
+DTKCORE_EXPORT QDataStream& operator >> (QDataStream&, dtkCoreParameterInListStringList&);
+
+DTKCORE_EXPORT QDebug operator << (QDebug, dtkCoreParameterInListStringList);
+
+
 // ///////////////////////////////////////////////////////////////////
 // dtkCoreParameterRange declaration
 // ///////////////////////////////////////////////////////////////////
@@ -402,6 +466,7 @@ public:
     using self_type = dtkCoreParameterRange<value_type>;
     using base_type = dtkCoreParameterBase<self_type>;
 
+    using dtkCoreParameter::castToTop;
     using dtkCoreParameter::documentation;
     using dtkCoreParameter::setDocumentation;
     using dtkCoreParameter::label;
@@ -484,6 +549,8 @@ namespace dtk {
     using d_string = dtkCoreParameterSimple<QString>;
 
     using d_inliststring = dtkCoreParameterInList<QString>;
+    using d_inliststringlist = dtkCoreParameterInListStringList;
+
     using d_inlistreal = dtkCoreParameterInList<double>;
     using d_inlistint = dtkCoreParameterInList<qlonglong>;
 
@@ -507,6 +574,7 @@ DTK_DECLARE_PARAMETER(dtk::d_bool);
 DTK_DECLARE_PARAMETER(dtk::d_string);
 
 DTK_DECLARE_PARAMETER(dtk::d_inliststring);
+DTK_DECLARE_PARAMETER(dtk::d_inliststringlist);
 DTK_DECLARE_PARAMETER(dtk::d_inlistreal);
 DTK_DECLARE_PARAMETER(dtk::d_inlistint);