+++ tst/dtkCore/dtkCoreParameterTest.cpp
@@ -54,6 +54,7 @@ void dtkCoreParameterTestCase::testRegistration(void)
     QVERIFY(QMetaType::type("dtk::d_range_real") != QMetaType::UnknownType);
     QVERIFY(QMetaType::type("dtk::d_inliststring") != QMetaType::UnknownType);
     QVERIFY(QMetaType::type("dtkCoreParameterSimple<std::string>") != QMetaType::UnknownType);
+    QVERIFY(QMetaType::type("dtk::d_inliststringlist") != QMetaType::UnknownType);
 }
 
 void dtkCoreParameterTestCase::testCreation(void)
@@ -953,6 +954,25 @@ void dtkCoreParameterTestCase::testRange(void)
     range_r2.setValue( {a, b } );
 }
 
+void dtkCoreParameterTestCase::testStringList(void)
+{
+    dtk::d_inliststringlist param_list("name", {"aa", "bb"}, {"aa", "bb", "cc", "dd"});
+    QVERIFY(param_list.label() == "name");
+
+    QCOMPARE(param_list.value().size(), 2);
+    QCOMPARE(param_list.values().size(), 4);
+
+    param_list.addValue("zz");
+    QCOMPARE(param_list.value().size(), 2);
+    QCOMPARE(param_list.values().size(), 5);
+
+
+    param_list.removeValue("aa");
+    QCOMPARE(param_list.value().size(), 1);
+    QCOMPARE(param_list.values().size(), 4);
+
+}
+
 void dtkCoreParameterTestCase::testFile(void)
 {
     dtk::d_path source("file", "/pim/pam/poum/pipo.jpg", {"*.jpg", "*.png"},  "File parameter example");
@@ -1016,12 +1036,57 @@ void dtkCoreParameterTestCase::testReadParameters(void)
         QCOMPARE(res["dif"]->documentation() , QString("diffusion Flag"));
         QCOMPARE(dtk::d_bool(res["dif"]->variant()).value() , false);
         QCOMPARE(dtk::d_real(res["toto"]->variant()).value() , 3.1415);
+
+        res = dtk::core::readParameters(json_file);
+        QCOMPARE(res.count(), 6);
+
+        QCOMPARE(res["toto"]->uid(), QString("toto"));
+
+        QCOMPARE(res["hyp"]->label(), QString("Porosity Model"));
+        QCOMPARE(dtk::d_int(res["hyp"]->variant()).value(), 2);
+
+        QCOMPARE(res["dif"]->documentation(), QString("diffusion Flag"));
+        QCOMPARE(dtk::d_bool(res["dif"]->variant()).value(), false);
+        QCOMPARE(dtk::d_real(res["toto"]->variant()).value(), 3.1415);
     }
 
     {
         auto resbad = dtk::core::readParameters(json_bad_file);
         QCOMPARE(resbad.count() , 0);
     }
+}
+
+void dtkCoreParameterTestCase::testReadParametersResources(void)
+{
+    QString json_file(QFINDTESTDATA(":/params.json"));
+
+    {
+        auto res = dtk::core::readParameters(json_file);
+        QCOMPARE(res.count(), 6);
+
+        QCOMPARE(res["toto"]->uid(), QString("toto"));
+
+        QCOMPARE(res["hyp"]->label(), QString("Porosity Model"));
+        QCOMPARE(dtk::d_int(res["hyp"]->variant()).value(), 2);
+
+        QCOMPARE(res["dif"]->documentation(), QString("diffusion Flag"));
+        QCOMPARE(dtk::d_bool(res["dif"]->variant()).value(), false);
+        QCOMPARE(dtk::d_real(res["toto"]->variant()).value(), 3.1415);
+
+         res = dtk::core::readParameters(json_file);
+        QCOMPARE(res.count(), 6);
+
+        QCOMPARE(res["toto"]->uid(), QString("toto"));
+
+        QCOMPARE(res["hyp"]->label(), QString("Porosity Model"));
+        QCOMPARE(dtk::d_int(res["hyp"]->variant()).value(), 2);
+
+        QCOMPARE(res["dif"]->documentation(), QString("diffusion Flag"));
+        QCOMPARE(dtk::d_bool(res["dif"]->variant()).value(), false);
+        QCOMPARE(dtk::d_real(res["toto"]->variant()).value(), 3.1415);
+
+    }
+
 }
 
 void dtkCoreParameterTestCase::testToVariantHash(void)