+++ tst/dtkCore/dtkCoreParameterTest.h
@@ -32,8 +32,10 @@ private slots:
     void testConnection(void);
     void testText(void);
     void testRange(void);
+    void testStringList(void);
     void testFile(void);
     void testReadParameters(void);
+    void testReadParametersResources(void);
     void testToVariantHash(void);
 
 private slots: